backend/pyproject.toml
[project]
name = "mina-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aio-pika>=9.5.8",
    "alembic>=1.18.1",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "clamd>=1.0.2",
    "fastapi-mail>=1.6.1",
    "fastapi[standard]>=0.128.0",
    "greenlet>=3.3.1",
    "minio>=7.2.20",
    "passlib>=1.7.4",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.22",
    "redis[hiredis]>=7.2.0",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "coverage>=7.13.3",
    "faker>=40.4.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
    "testcontainers[minio]>=4.14.1",
]

[tool.black]
line-length = 81
target-version = ["py312"]
skip-string-normalization = true
skip-magic-trailing-comma = true
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.mypy]
exclude = "^(venv/|\\.venv/|build/|dist/|migrations/)"
incremental = true
cache_dir = ".mypy_cache"
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/conftest.py",
    "*/migrations/*",
    "*/__init__.py",
    "*/site-packages/*"
]
